Rename face-api module to be just api.js
