+++ frontend/services/__tests__/api.test.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import MockAdapter from 'axios-mock-adapter';
-import * as api from '../face-api';
+import * as api from '../api';
 
 describe('API', () => {
   let mock;
